requirements.txt
Flask==0.12
mysql-connector==2.1.6
Flask==1.0.2